Move internal methods and classes to Internal namespace
