+++ tests/ColorThiefTest.php
@@ -18,7 +18,6 @@ use ColorThief\ColorThief;
 use ColorThief\Exception\InvalidArgumentException;
 use ColorThief\Exception\NotSupportedException;
 use ColorThief\Image\Adapter\AdapterInterface;
-use ColorThief\VBox;
 use PHPUnit\Framework\Attributes\DataProvider;
 
 class ColorThiefTest extends \PHPUnit\Framework\TestCase
@@ -137,24 +136,6 @@ class ColorThiefTest extends \PHPUnit\Framework\TestCase
                 ],
             ],
             */
-        ];
-    }
-
-    public static function provide8bitsColorIndex(): array
-    {
-        return [
-            [0, 0, 0, 0],
-            [120, 120, 120, 7895160],
-            [255, 255, 255, 16777215],
-        ];
-    }
-
-    public static function provide5bitsColorIndex(): array
-    {
-        return [
-            [0,     0,   0,  0b000000000000000],
-            [120, 120, 120,  0b011110111101111],
-            [255, 255, 255,  0b111111111111111],
         ];
     }
 
@@ -329,132 +310,5 @@ class ColorThiefTest extends \PHPUnit\Framework\TestCase
         $this->expectExceptionMessage('blank or transparent image');
 
         ColorThief::getPalette(__DIR__.'/images/blank.png');
-    }
-
-    #[DataProvider('provide8bitsColorIndex')]
-    public function testGetColorIndex8bits(int $r, int $g, int $b, int $index): void
-    {
-        $this->assertSame(
-            $index,
-            ColorThief::getColorIndex($r, $g, $b, 8)
-        );
-    }
-
-    #[DataProvider('provide5bitsColorIndex')]
-    public function testGetColorIndex5bits(int $r, int $g, int $b, int $index): void
-    {
-        $this->assertSame(
-            $index,
-            ColorThief::getColorIndex($r, $g, $b)
-        );
-    }
-
-    /**
-     * Tests RGB values are the same after converting them back from combined bucket index to RGB bucket values.
-     */
-    #[DataProvider('provide5bitsColorIndex')]
-    public function testGetColorsFromIndex5bits(int $r, int $g, int $b, int $index): void
-    {
-        $rgbBuckets = [$r >> ColorThief::RSHIFT, $g >> ColorThief::RSHIFT, $b >> ColorThief::RSHIFT];
-        $this->assertSame(
-            [$rgbBuckets[0], $rgbBuckets[1], $rgbBuckets[2]],
-            ColorThief::getColorsFromIndex($index, ColorThief::SIGBITS)
-        );
-    }
-
-    #[DataProvider('provide8bitsColorIndex')]
-    public function testGetColorsFromIndex8bits(int $r, int $g, int $b, int $index): void
-    {
-        $this->assertSame(
-            [$r, $g, $b],
-            ColorThief::getColorsFromIndex($index)
-        );
-    }
-
-    public function testVboxFromPixels(): void
-    {
-        $method = new \ReflectionMethod(ColorThief::class, 'vboxFromHistogram');
-
-        // [[229, 210, 51], [133, 24, 135], [216, 235, 108], [132, 25, 134], [223, 46, 29],
-        // [135, 28, 132], [233, 133, 213], [225, 212, 48]]
-        // $pixels = array(15061555, 8722567, 14216044, 8657286, 14626333, 8854660, 15304149, 14799920);
-
-        $histo = [
-            29510 => 2,
-            16496 => 3,
-            28589 => 1,
-            27811 => 1,
-            30234 => 1,
-        ];
-
-        $result = $method->invoke(null, $histo);
-
-        $this->assertInstanceOf(VBox::class, $result);
-        $this->assertSame($histo, $result->histo);
-        $this->assertSame(16, $result->r1);
-        $this->assertSame(29, $result->r2);
-        $this->assertSame(3, $result->g1);
-        $this->assertSame(29, $result->g2);
-        $this->assertSame(3, $result->b1);
-        $this->assertSame(26, $result->b2);
-    }
-
-    /**
-     * Tests min and max RGB values are equal if there is only one color in the histogram (PR #41).
-     */
-    public function testVboxFromSingleColorHistogram(): void
-    {
-        $method = new \ReflectionMethod(ColorThief::class, 'vboxFromHistogram');
-
-        $histo = [
-            26756 => 120000,
-        ];
-
-        $result = $method->invoke(null, $histo);
-
-        $this->assertInstanceOf(VBox::class, $result);
-        $this->assertSame($histo, $result->histo);
-        $this->assertSame($result->r1, $result->r2);
-        $this->assertSame($result->g1, $result->g2);
-        $this->assertSame($result->b1, $result->b2);
-        $this->assertSame(1, $result->volume());
-    }
-
-    public function testDoCutLeftLeatherThanRight(): void
-    {
-        $method = new \ReflectionMethod(ColorThief::class, 'doCut');
-
-        // $left <= $right
-        $result = $method->invoke(
-            null,
-            'g',
-            new VBox(0, 20, 0, 31, 0, 31, []),
-            [38, 149, 556, 1222, 1830, 2656, 3638, 4744, 6039, 7412, 9039, 10686, 12244, 13715, 15091, 16355, 17599, 18768, 19771,
-                20925, 22257, 24094, 25782, 27585, 28796, 29794, 30258, 30290, 30298, 30301, 30301, 30301, ],
-            30301,
-            [30263, 30152, 29745, 29079, 28471, 27645, 26663, 25557, 24262, 22889, 21262, 19615, 18057, 16586, 15210, 13946,
-                12702, 11533, 10530, 9376, 8044, 6207, 4519, 2716, 1505, 507, 43, 11, 3, 0, 0, 0, ]
-        );
-
-        $this->assertEquals(new VBox(0, 20, 0, 23, 0, 31, []), $result[0]);
-        $this->assertEquals(new VBox(0, 20, 24, 31, 0, 31, []), $result[1]);
-    }
-
-    public function testDoCutLeftGreaterThanRight(): void
-    {
-        $method = new \ReflectionMethod(ColorThief::class, 'doCut');
-
-        // $left > $right
-        $result = $method->invoke(
-            null,
-            'g',
-            new VBox(0, 13, 0, 17, 0, 10, []),
-            [38, 149, 512, 1151, 1741, 2554, 3530, 4624, 5899, 7247, 8788, 10261, 11645, 12906, 13969, 14871, 15654, 16329],
-            16329,
-            [16291, 16180, 15817, 15178, 14588, 13775, 12799, 11705, 10430, 9082, 7541, 6068, 4684, 3423, 2360, 1458, 675, 0]
-        );
-
-        $this->assertEquals(new VBox(0, 13, 0, 4, 0, 10, []), $result[0]);
-        $this->assertEquals(new VBox(0, 13, 5, 17, 0, 10, []), $result[1]);
     }
 }
